upgrade setting
